fix - Move widgets to main app
